in/out integration tests

+++ crud/src/main/java/com/example/crud/repository/CrudRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.crud.model.CrudModel;
 
-public interface CrudRepository extends JpaRepository<CrudModel, Short> {
+public interface CrudRepository extends JpaRepository<CrudModel, Integer> {
 
 }
